# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KANIDA.AI API
  version: 0.1.0
  description: 'The KANIDA.AI product contract — the seam between the engine track and the app track.


    Product laws that are enforced *in the schemas*, not merely documented:

    - **Expectancy is the hero metric**; `win_rate_pct` is optional and never the headline.

    - **Every return is paired with its drawdown** — a performance block cannot exist without `max_drawdown_pct`
    and `current_drawdown_pct`.

    - **No return promises and no target prices** — there is deliberately no target, projection or expected-return
    field anywhere in this contract.

    - **Every number carries n + date range + data source + cost convention** via `Provenance`; `sample_flag`
    is derived (`greyed` below n=20, `flagged` below n=50).

    - **Point-in-time is law** — every `Provenance` carries an `as_of`.

    - **Losers first** — ledgers are named `ledger_losers_first` and sorted ascending by net P&L.

    - **Backtests are labelled** ''Simulated · Not traded · Not PaRRVA-verified''.

    - **The LLM never calculates** — narrative text authored by a model contains no literal numerals,
    only `{{fact:<id>}}` references resolved against `facts[]`.'
paths:
  /api/pathfinder/experiment/{experiment_id}:
    get:
      tags:
      - Pathfinder
      summary: One experiment's full journey
      description: '**S2 (`KANIDA_PATHFINDER_SOURCE=research`)**: the in-app `ExperimentRecord` — versions
        (v1, v2… with what changed, why, the trial count and the expectation FROZEN when each opened),
        every counted trial, every period with its frozen grading rule, forward result, expected-vs-actual
        and learning, the change-log, the post-mortem when buried, and the graduation PROPOSAL when one
        exists (human-gated). Constituent names are withheld pending RA review.


        Other sources serve the P0/P1 `ExperimentDetail`.'
      operationId: get_experiment_api_pathfinder_experiment__experiment_id__get
      parameters:
      - name: experiment_id
        in: path
        required: true
        schema:
          type: string
          title: Experiment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ExperimentRecord'
                - $ref: '#/components/schemas/ExperimentDetail'
                title: Response Get Experiment Api Pathfinder Experiment  Experiment Id  Get
        '400':
          description: Malformed request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (details are not disclosed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/pathfinder/experiments:
    get:
      tags:
      - Pathfinder
      summary: The edge-discovery pipeline
      description: 'Every experiment across the lifecycle. Dead experiments are listed, not hidden.


        **S2 (`KANIDA_PATHFINDER_SOURCE=research`)** serves the experiment REGISTRY as `ExperimentsResponse`:
        every experiment''s PUBLIC card (theme + evidence + the seven-line story — never a constituent,
        an entry, a target or a stop), the S1 findings the gate declined with their trial counts (`not_opened`),
        and the running experiment scoreboard (Right · Wrong · Inconclusive · n over graded periods, void
        apart, forward / backfilled split). Losers first: buried, then testing, proposed last.


        Other sources serve the P0/P1 `ExperimentListResponse`.'
      operationId: get_experiments_api_pathfinder_experiments_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ExperimentStatus'
          - type: 'null'
          description: 'P0/P1 only: filter by lifecycle status. Omit for all.'
          title: Status
        description: 'P0/P1 only: filter by lifecycle status. Omit for all.'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ExperimentsResponse'
                - $ref: '#/components/schemas/ExperimentListResponse'
                title: Response Get Experiments Api Pathfinder Experiments Get
        '400':
          description: Malformed request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (details are not disclosed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/pathfinder/feed:
    get:
      tags:
      - Pathfinder
      summary: The clarity-first research feed for one close (S1)
      description: 'The after-close edition: findings ranked by usefulness, **clarity first** — `what_matters_now`
        holds the first two or three, `discoveries` the rest. There is **no minimum count and no padding**:
        an edition with four findings is a valid edition.


        Every finding was **computed before it was written** by a template from the question library and
        carries: a digit-free `narrative` whose numbers are `{{fact:<id>}}` references into `facts[]`;
        full `provenance` (evidence level, n, period, regime, comparison group, cost hurdle); a `grading_rule`
        **frozen at publication**; and its `grading` state (pending, or graded Right / Wrong / Inconclusive
        with the realised facts). `scoreboard` is the running public record with n.


        `date` selects an edition (YYYY-MM-DD); omit it for the latest. Research items only — no entry,
        target, stop or execution is ever implied.'
      operationId: get_feed_api_pathfinder_feed_get
      parameters:
      - name: date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Edition close date, YYYY-MM-DD. Omit for the latest.
          title: Date
        description: Edition close date, YYYY-MM-DD. Omit for the latest.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
        '400':
          description: Malformed request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (details are not disclosed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/pathfinder/learnings:
    get:
      tags:
      - Pathfinder
      summary: What Pathfinder has learned, and what it is testing next
      description: '`learned[]` is what the agent now believes, each tied to the experiments and evidence
        that produced it and tagged with its learning level (L1 evidence, L2 parameter, L3 strategy —
        L4 Constitution changes are human-only and never authored here). `testing_next[]` is what it decided
        to test next and why now, including the ones that are blocked and why.'
      operationId: get_learnings_api_pathfinder_learnings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningsResponse'
        '400':
          description: Malformed request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (details are not disclosed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pathfinder/loop:
    get:
      tags:
      - Pathfinder
      summary: The live research loop, as a story
      description: 'One cycle of the loop: what the deterministic observer noticed, the hypothesis it
        justified waking the LLM for, the virtual experiment, what happened, what was learned and changed,
        and what is next — plus the supporting `facts` every narrative sentence references.


        **Every number in this payload originates in `facts[]`, computed by a deterministic engine.**
        LLM-authored story text contains no literal numerals; it carries `{{fact:<id>}}` references the
        client resolves against `facts[]`.'
      operationId: get_loop_api_pathfinder_loop_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoopResponse'
        '400':
          description: Malformed request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No such resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (details are not disclosed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    Direction:
      type: string
      enum:
      - long
      - short
      title: Direction
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      additionalProperties: false
      type: object
      required:
      - error
      title: ErrorResponse
...
    ExperimentDetail:
      properties:
        id:
          type: string
          pattern: ^exp_[a-z0-9_]+$
          title: Id
        hypothesis:
          type: string
          title: Hypothesis
        status:
          $ref: '#/components/schemas/ExperimentStatus'
        opened_at:
          type: string
          format: date-time
          title: Opened At
        strategy_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Strategy Version
        constitution_version:
          type: string
          title: Constitution Version
        historical_return:
          anyOf:
          - $ref: '#/components/schemas/PerformanceBlock'
          - type: 'null'
          description: Strategy-replay over the discovery window. Null while queued.
        virtual_return:
          anyOf:
          - $ref: '#/components/schemas/PerformanceBlock'
          - type: 'null'
          description: The forward virtual book. Null until the first closed trade.
        occurrences:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Occurrences
        n:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: N
        sample_flag:
          $ref: '#/components/schemas/SampleFlag'
          default: unknown
        spark:
          anyOf:
          - $ref: '#/components/schemas/Spark'
          - type: 'null'
        as_of:
          type: string
          format: date-time
          title: As Of
        question:
          type: string
          title: Question
          description: The question the agent asked itself.
        rationale:
          type: string
          title: Rationale
        rulebook:
          $ref: '#/components/schemas/Rulebook'
        stage:
          $ref: '#/components/schemas/LoopStage'
        story:
          items:
            $ref: '#/components/schemas/StoryLine'
          type: array
          minItems: 1
          title: Story
          description: The loop story, in beat order.
        facts:
          items:
            $ref: '#/components/schemas/Fact'
          type: array
          title: Facts
          description: Every fact referenced anywhere in this payload.
        evidence:
          items:
            $ref: '#/components/schemas/Evidence'
          type: array
          title: Evidence
        virtual_book:
          anyOf:
          - $ref: '#/components/schemas/VirtualBook'
          - type: 'null'
        change_log:
          items:
            $ref: '#/components/schemas/ChangeLogEntry'
          type: array
          title: Change Log
          description: Append-only L1–L3 history (an L4 change is recorded, never authored here).
        post_mortem:
          anyOf:
          - $ref: '#/components/schemas/PostMortem'
          - type: 'null'
        triggers:
          items:
            $ref: '#/components/schemas/Trigger'
          type: array
          title: Triggers
        llm_usage:
          anyOf:
          - $ref: '#/components/schemas/LlmUsage'
          - type: 'null'
        next_review:
          anyOf:
          - $ref: '#/components/schemas/Trigger'
          - type: 'null'
          description: The condition that will next wake the LLM for this experiment.
        disclosure:
          type: string
          title: Disclosure
          default: Pathfinder is a research lab. Everything here is a hypothesis under test with virtual
            money, not advice and not a recommendation. No returns are promised. Kanida never places an
            order on your behalf.
      additionalProperties: false
      type: object
      required:
      - id
      - hypothesis
      - status
      - opened_at
      - constitution_version
      - as_of
      - question
      - rationale
      - rulebook
      - stage
      - story
      - facts
      title: ExperimentDetail
      description: The full journey — this is the screen the whole product is about.
...
    ExperimentRecord:
      properties:
        id:
          type: string
          pattern: ^exp_[a-z0-9_]+$
          title: Id
        opened_edition:
          type: string
          format: date
          title: Opened Edition
        news_edition:
          type: string
          format: date
          title: News Edition
          description: The edition this card's story was written on.
        state:
          $ref: '#/components/schemas/ExperimentState'
        family:
          type: string
          title: Family
        theme:
          type: string
          title: Theme
          description: Sector / theme in words. Never a name, never an instruction.
        source_finding_id:
          type: string
          pattern: ^fnd_[a-z0-9_]+$
          title: Source Finding Id
        evidence:
          $ref: '#/components/schemas/FindingProvenance'
        story:
          items:
            $ref: '#/components/schemas/ExperimentStoryLine'
          type: array
          maxItems: 7
          minItems: 7
          title: Story
        facts:
          items:
            $ref: '#/components/schemas/Fact'
          type: array
          minItems: 1
          title: Facts
        versions_count:
          type: integer
          minimum: 1.0
          title: Versions Count
        trials_total:
          type: integer
          minimum: 1.0
          title: Trials Total
          description: Every variant ever evaluated for this experiment.
        periods_graded:
          type: integer
          minimum: 0.0
          title: Periods Graded
        score:
          $ref: '#/components/schemas/ExperimentScoreCounts'
        latest_comparison:
          $ref: '#/components/schemas/ComparisonCategory'
        backfilled:
          type: boolean
          title: Backfilled
          description: 'Whether the edition this card was written on (`news_edition`) was generated after
            its session date. Per edition, like every S1 card (audit finding 1): a card can be forward
            while the experiment was opened in a backfill.'
        opened_backfilled:
          type: boolean
          title: Opened Backfilled
          description: Whether the experiment's OPENING edition was a backfill.
        record_label:
          type: string
          title: Record Label
        llm_provider:
          type: string
          title: Llm Provider
          default: none
        disclosure:
          type: string
          title: Disclosure
          default: Research experiment with virtual money — research validation, not advice and not a
            recommendation. Nothing here tells you when to enter, exit, or where to place a stop; constituent
            names appear only inside the app after Research Analyst review. Kanida never places an order
            on your behalf.
        family_trials_all_time:
          type: integer
          minimum: 0.0
          title: Family Trials All Time
          description: 'S2 re-audit N4: every trial this experiment''s FAMILY has ever had as of this
            card''s edition — across every finding, retry and revision — the count the family-wise significance
            bar divides by. Never restarts.'
          default: 0
        question:
          type: string
          title: Question
        current_rule_text:
          type: string
          title: Current Rule Text
        direction:
          $ref: '#/components/schemas/Direction'
        versions:
          items:
            $ref: '#/components/schemas/VersionView'
          type: array
          minItems: 1
          title: Versions
        trials:
          items:
            $ref: '#/components/schemas/TrialView'
          type: array
          title: Trials
        basket:
          $ref: '#/components/schemas/BasketView'
        worth_testing_gates:
          items:
            $ref: '#/components/schemas/GateView'
          type: array
          title: Worth Testing Gates
        change_log:
          items:
            $ref: '#/components/schemas/ChangeLogEntry'
          type: array
          title: Change Log
        post_mortem:
          anyOf:
          - $ref: '#/components/schemas/PostMortem'
          - type: 'null'
        proposal:
          anyOf:
          - $ref: '#/components/schemas/ProposalView'
          - type: 'null'
        constitution_version:
          type: string
          title: Constitution Version
      additionalProperties: false
      type: object
      required:
      - id
      - opened_edition
      - news_edition
      - state
      - family
      - theme
      - source_finding_id
      - evidence
      - story
      - facts
      - versions_count
      - trials_total
      - periods_graded
      - score
      - latest_comparison
      - backfilled
      - opened_backfilled
      - record_label
      - question
      - current_rule_text
      - direction
      - versions
      - trials
      - basket
      - worth_testing_gates
      - constitution_version
      title: ExperimentRecord
      description: 'The in-app record: versions, trials, periods, expected vs actual, learning, proposal,
        post-mortem.'
...
    ExperimentState:
      type: string
      enum:
      - testing
      - buried
      - proposed
      title: ExperimentState
    ExperimentStatus:
      type: string
      enum:
      - queued
      - testing
      - validating
      - promising
      - promoted
      - died
      title: ExperimentStatus
      description: The Pathfinder lifecycle. `died` is a first-class, published outcome.
...
    ExperimentsResponse:
      properties:
        as_of:
          type: string
          format: date
          title: As Of
        count:
          type: integer
          minimum: 0.0
          title: Count
        items:
          items:
            $ref: '#/components/schemas/ExperimentCard'
          type: array
          title: Items
        not_opened:
          items:
            $ref: '#/components/schemas/RejectedCandidate'
          type: array
          title: Not Opened
        scoreboard:
          $ref: '#/components/schemas/ExperimentScoreboard'
        llm_provider:
          type: string
          title: Llm Provider
          default: none
        engine_version:
          type: string
          title: Engine Version
        disclosure:
          type: string
          title: Disclosure
          default: Research experiment with virtual money — research validation, not advice and not a
            recommendation. Nothing here tells you when to enter, exit, or where to place a stop; constituent
            names appear only inside the app after Research Analyst review. Kanida never places an order
            on your behalf.
      additionalProperties: false
      type: object
      required:
      - as_of
      - count
      - items
      - scoreboard
      - engine_version
      title: ExperimentsResponse
      description: 'GET /api/pathfinder/experiments (S2). Losers first: buried, then testing, proposed
        last.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...